pkg/trace/agent: simplify channels (#3868)
This change simplifies the way the channels are defined in the Agent
structure, by putting the responsibility of receiving from the receiver
channel into its maker. It makes more sense from an API standpoint where
a channel is passed to a constructor, rather than accessing that
structure's field directly (Receiver.Out).