[tagger] change prefix for container runtimes and k8s pods (#3821)
* [tagger] change prefix for container runtimes and k8s pods

Specifically, removes the runtime name for a container and simply
use "container_id://". As for pods, kubernetes_pod:// becomes
kubernetes_pod_uid://. The goal of this change is to make it simpler
to use the tagger from other agents (trace in this case) w/o having to
determine the runtime in use.