Handle 32bits overflow on 64bits system
Some value in /proc/diskstats overflow at maxUInt32 even when running on
a 64bit system.

We now detect if the value overflowed at the maxUInt32 mark or
maxUInt64. This fixes spikes in metrics when the values wrap back to 0.

This also fixes the values being offset by 1 when they were wrapping.