[kitchen] Test the branch's install script in kitchen
Until now, the kitchen tests used the install script on master
to do their tests. This is not practical if someone wants to
change the install script (as the person would have to wait
until their PR is merged before being able to test with kitchen).
The kitchen configuration has been changed to use the current
commit's install script.