[rtloader] Fix encoding of yaml.dump in Python 3
In Python 3, yaml.dump has 'utf-8' as its default encoding, which
results in a bytes string being returned (b'<yaml>').
When we convert it to a go string, we get b'<yaml>' which is
not proper yaml, thus causing a yaml unmarshal error in Go.
This error prevents the "Add a check" list from being displayed in
the GUI page where checks are managed.

The solution used here is to set the encoding to None: that way yaml.dump
returns a simple string. An alternative solution could be to keep
the default encoding, and decode the string before unmarshalling it if
Python 3 is loaded.