Remove yum warnings while removing the Agent package. (#3781)
Since 6.12.0, the prerm and preinst scripts take care of removing the embedded integrations,
however, most of those files are installed by yum with the package.
A package manager tries to remove all the files that it has installed while installing a pkg, apt
doesn't seem to complain when some of these files are missing, but yum/rpm does, and, as
we're removing files on our own in scripts, it results with yum throwing a lot of warnings.

In order to avoid this, I've added a step in the finalize step of the package creation: it creates a
file listing every files in /opt/datadog-agent/embedded/lib/python2.7/site-packages. This file
is later used to compare what has been installed by the package and what is actually on the
filesystem, to remove only what was not installed by the package.