[logs] Release file descriptor on file rotation when tailer is stuck (#4055)
* [logs] Rename tl into tailer in tailer_test.go

* [logs] Close tailer on file rotation even when stuck

Currently if the sender could not send the logs because it can't establish a
connection to the logs intake, the whole pipeline is stuck. Thus when there is
a file rotation the previous tailer is never close and it keeps a file
descriptor causing a leak.

This commit makes possible to close a tailer even if it is stuck thanks to the
use of a context.