pkg/trace/obfuscate: ensure double-quoted strings after assignments are obfuscated
This change ensures that double-quoted strings after assignments get
obfuscated. In general, double-quoted strings can denote identifiers in
some cases, e.g.:

> If the ANSI_QUOTES SQL mode is enabled, string literals can be quoted only within single quotation marks because a string quoted within double quotation marks is interpreted as an identifier.
Ref: https://dev.mysql.com/doc/refman/8.0/en/string-literals.html

Using double quotes for strings is generally considered bad practice but
in some cases is omitted. It should be safe, until proven wrong, to
consider double quoted strings following an assignments to be a token
that needs to be obfuscated.