[Scheduler] SendNow method to run a collector on demand (#4166)
Adds a SendNow(collector) method to Scheduler that triggers a Send
call on the given collector and restarts its regular collection timer. To do so,
the time.Ticker has been replaced by a time.Timer, that can be reset.
This has the implication that we now must reset it each time after it fires.

Adds a tests that verifies both ways to call Send: at regular intervals
and via SendNow.

Also, it fixes the Stop test, which was trying to add inexistent collectors, and
makes tests faster by disabling the firstRun code that ran the host collector.