pkg/trace/api: avoid misleading rate limiting rate when inactive
When the X-Datadog-Trace-Count header is missing, the ratelimiter is
technically inactive. Never the less, its target rate continues to be
updated by monitoring the process's CPU and memory usage and was
incorrectly appended to span, resulting in skewed stats.

This change ensures that when the rate limiter has seen no traces, it
reports a permission rate of 100%.