Keep track of connection type in conntracker.go (#4256)
Include L4 proto in the connection key used by conntracker.go

the old key for each conntrack entry was (sourceIP, sourcePort). Because it's possible to use have UDP and TCP connections have the same sourceIP/sourcePort, a bunch of TCP connections were incorrectly . attached to UDP translations.

Also includes an integration test which uses IP tables to simulate a DNAT-ed connection. The integration test needs to be run as root, and will not run otherwise.