[gitlab] Fix the logic that checks existence of deb pkgs, for Agent 7
Now that we release both Agent 6 and 7, the script that checks the
existence of already-released deb packages matches all deb pkgs
present locally.

Since we run this check on both `deploy_deb-6` and `deploy_deb-7`,
make sure the check in these jobs only checks respectively the v6
and v7 packages, otherwise one of them may fail on a race condition.