logs: store at creation if a input File has been discovered with a wild tag.
The searchFiles method which is looking for file to tail in a directory
is aware if a file has been discovered through a wild card pattern. If
so, this change use a changed constructor of File to immediatly store if
it's the case, instead of trying to infer it later (and fail in some cases).