Fixes agent binary to not be dependent onthe VC9 runtime on A7 (#4339)
(when we don't need the VC9 runtime)

Makes a conditional compile of a different agent manifest.
The existing agent manifest specifies that the agent binary requires the specific VC9 runtime that we ship with Python2. For agent 7 (or, specifically builds that don't have py2), we also don't ship that
runtime. Then the agent binary couldn't load due to an unmet dependency.

Creates a new manifest, one for Py2 and one for Py3. Compiles using the correct manifest.