[kitchen] Fix install script kitchen test not finding apt/yum repo
We modified our install script to make it work for any release branch
and component.
However this broke the kitchen tests, which use sed to replace parts
of the install script to make it work with the testing repository.

A more durable solution would be to use the new environment variables
provided (which would require more work on the chef recipe used in the
kitchen test), but this should be enough to unblock the pipeline for now.