Force quit after the grace period while closing the logs agent.
We use different stoppers (serial/parallel) to stop everything in the
logs agent in order to not lose any line of log.

However, in some circumstances (reproduced only on Windows) when no network
is available while closing the Agent, the stoppers never succesfully complete.
In this case, the Agent is stopping and there is no network: we won't be able
to send the logs anyway.

This PR adds a small timer to give some last seconds to the stoppers and
then finally force close the Log Agent.