Re-order checks
See: https://github.com/DataDog/dd-agent/issues/3860

The default systemd socket is a private API not intended for external use. For this reason, processes must be run as root to access them, and therefore this is typically very unlikely to work as the agent does not run as root by default. We re-order the checks to try dbus first which is far more likely to succeed.

This has the benefit of not causing extensive logging is syslog for failed connections due to insufficient permissions.


Co-authored-by: Alistair Lynn <arplynn@gmail.com>