Make the agent retrieve the cloud instance-id from the kubelet
When the agent is started with `hostNetwork: true`, it is able to find the AWS/GCE instance-id
from the metadata agent (169.254.169.254).
When the agent is started with `hostNetwork: false`, that metadata agent should become unreachable for security reasons.
In that case, the current kubelet hostname provider tries to concatenate the node name with the cluster name.
This change makes the kubelet hostname provider first try to get the instance-id from the kubelet.