Use erb template vars in SysVInit scripts
We use erb templates for our service scripts, and we provide
two vars (etc_dir and install_dir), which are useful if we
want to change the installation or config directories.
However, we didn't use them in the SysVInit scripts (the
etc and install directories were hardcoded in the service files).