Use erb template vars in SysVInit scripts (#4457)
Make use of the etc_dir and install_dir template variables 
provided to the service template files.

We use erb templates for our service scripts, and we
provide two vars (etc_dir and install_dir), which are useful
if we want to change the installation or config directories
and still have working scripts. However, we didn't use them
in the SysVInit scripts (the etc and install directories were
hardcoded in the service files).