[kitchen] Upgrade chef-datadog cookbook to 4.0 + cleanup (#4515)
Uses version 4.0 (has yet to be released) of the chef-datadog cookbook.

Changes the apt component where the kitchen artifacts are uploaded, since chef-datadog now expects Agent 6 packages to be in the 6 component, and Agent 7 packages to be in the 7 component.

Also changes the yum branch where the kitchen artifacts are uploaded to match the apt structure (Agent 6 packages are now in pipeline-<ID>/6, and Agent 7 packages in pipeline-<ID>/7).

Small fixes to the cleanup task (A6 and A7 boxes were tagged with the same tag, so both A6 and A7 Azure cleanup jobs would try to delete all kitchen VMs instead of only the VMs of the corresponding Agent version).

Cleanup of the kitchen cookbooks: makes a kitchen cookbook only use its own variables, and not variables from another cookbook (makes reading and understanding the cookbooks easier, as well as prevent unexpected behavior when changing a variable from a different cookbook). Also removes some unused or useless variables.