Allow the agent to run even if we can't create auth_token
On Windows the agent might not have the right to create the auth_token
for the API. If we exit for this, it triggers a rollback when installing
the agent (since the msi check that the service is able to start).

We now let the agent start so the install is successful even if the API
will not work.

This commit also edit some error logs in order to remove the 'token:'
which trigger the log stripper to remove the error message.