Make build work
A recent change tried to define asm_volatile_goto if it was not already
defined. The problem with this is subsequent header files (probably
compiler.h).

Bionic would not later try to redefine it, but older version of the
kernel would (e.g. my dev VM is running 4.4.0).

To get around this, include <linux/compiler.h> which should do the right
thing and make sure asm_volatile_goto is defined if necessary.