Make agent version functions take an explicit major version parameter  (#4667)
Make all invoke functions depending on the major version take the major version parameter explicitly, instead of relying on environment variables.

Omnibus builds are somewhat trickier, as we can't pass parameters to bundle exec omnibus, only environment variables, and the omnibus builds do need the major version, as they call invoke functions. The way we're doing is thus as follows:
- inv agent.omnibus-build takes an explicit --major-version parameter,
- It injects this parameter to the omnibus env vars as `MAJOR_VERSION,
the omnibus software definitions use the MAJOR_VERSION env var to call the other invoke functions (such as inv agent.build, inv dogstatsd.build, etc.) with the explicit --major-version parameter.

This PR creates a tasks/winbuildscripts folder containing the windows build scripts, as this will become the preferred way to package the build scripts once #4614 is merged (instead of baking the scripts in the build image directly, which is inconvenient).