pkg/trace/stats: do not ignore single span sublayer metrics (#4571)
No longer ignoring childless spans for apm sublayer metrics and stats to fix inaccuracies in how sublayer metrics are sometimes reported.

It looks like this was done as an optimization since it makes sense if the service is always sending a single span, then the sublayer metrics are actually irrelevant. But in cases where the service sometimes has child spans and sometimes doesn't, this can cause the metric to simply be inaccurate since it doesn't take all metrics into account.