pkg/trace/agent: use shared agent logger
This change removes the customer logger implementation and switches to
using the shared library logger. As a result, support for
`apm_config.log_throttling` is removed and instead we gain support for
full logging features such as `log_to_console` and `log_to_syslog`.

Log throttling is not something that users need to manually configure
and should be taken care of by developers to ensure that errors and
warnings never spam the user.

A new package `pkg/trace/logutil` has been added which exposes a
throttled logger helper allowing to limit how many times a certain
message appears during a time window.