pkg/trace/obfuscate: add a filter to extract table names from queries (#4399)
### Description

This change introduces a new tag called `sql.tables` which contains the tables that a query addresses as list of comma-separated values. To enable it, the feature flag `table_names` needs to be set using `DD_APM_FEATURES=table_names`. For example, given the query:

```
SELECT * FROM users JOIN friends
```

Will result in the tag `sql.tables` to contain the values `users,friends`

### Limitations

The current algorithm works agnostic of SQL engine. It simply takes the first word that follows any of the following keywords: `FROM`, `UPDATE`, `INTO` and `JOIN`. 

* Different SQL engines may have different syntax and could contain further keywords ([example](https://www.postgresql.org/docs/9.5/sql-select.html)) after these ones which are not table names. This case is not covered by the current implementation.

* If we later on decide to have the tracing clients mark the SQL engine as a tag, we could use it to further improve the processing.