[dogstatsd] Continue parsing packets with consecutive newlines (#4805)
When more than 1 consecutive newline was found, the logic would ignore
the rest of the packet. This worked fine so far probably because there
is no dogstatsd datagram out there that is badly formatted enough to
have 2 consecutive newlines in their contents.

That said, since the addition of the buffer of multiple packets, if a
small datagram has a trailing newline, the next small datagram can end
up appended to it with an additional newline separating the 2. In that
case, the parser will then see into 2 consecutive newlines. If it stops
parsing the packet at that point, it will have ignored all the messages
in the second datagram.

Fix this by continuing to parse the packet even after 2 or more
consecutive newlines are found.