Make the docker log processor use the global DockerUtil rather than its own docker client
The package `pkg/logs/input/docker` used to have its own docker client.
This could create some unwanted behavior in some edge cases.

One example is if the docker client used by the docker auto-discovery listener
is successfully connected to the docker daemon, it can start sending "new
docker source" and "new docker service" events to the log package.
If the docker client used by the log package is not connected to the docker
daemon, the "new docker source" and "new docker service" will not be properly
processed.

This change removes the docker client of the log package so that there's a
single docker client.
This will fix the behavior in the above described edge case.