Start system-probe unit with main Agent unit
- system-probe no longer needs to be enabled/started separately
- When the Agent is started, system-probe will start before it.
- system-probe won't start if config file is not present (see explanation below)
- If the system-probe doesn't start, Agent will still start without it.
- If Agent starts, system-probe also restarts (due to BindTo).

Ideally we don't want to run system-probe if it's not enabled, so in that
scenario we run nothing as root. This is hard to achieve without adding
some custom logic (that would have to run as root).

The middle ground I found was to use `ConditionPathExists=` to check for
the system-probe config file. Since, by default, that file doesn't exist
(we provide a .example file for users to copy), it means that by default
we will run nothing as root. If a user actually has a config file, they
probably want to run system-probe. It could happen that they create the
file but set `enabled: false` there, in that case we will start the
system-probe as root, do nothing and exit. I think it's still useful to do
so in this case, since it's most likely a configuration error and this
way we print a nice log saying "system probe not enabled. exiting."