Dockerfiles/agent: enable APM by default on Kubernetes
This change ensures that APM is automatically enabled in Kubernetes
deployments in order to simply onboarding and minimise support cases
about APM not working out of the box.

Additionally, the API rate limiter that was based on CPU and memory
usage thresholds is also disabled because it was another source of
endless escalations and complaints causing increase memory spikes (due
to the agent being blocked from taking full advantage of CPU power
because of this limitation) and lost data. Kubernetes already manages
container resource usage through its limits settings.