Build libbcc and system-probe is a dedicated build environment (#4871)
This change adds iovisor/bcc as a new runtime dependency of system-probe.
This is a dependency that comes from iovisor/gobpf which is already vendored in system-probe, but which is updated in this PR. (Strictly speeking, we are using the datadog fork of it).
iovisor/bcc requires a build environment which is more up-to-date than the centos:6 and debian:wheezy environment that are currently used to build the agent.
Previous attempts to build iovisor/bcc on the existing build environments failed because of the compiler being too old (lack of C++14 support), the Linux kernel header being too old and the glic headers being too old (netlink.h missing some constants)

So, the native library iovisor/bcc is built in a dedicated more up-to-date build environment created by DataDog/datadog-agent-buildimages#25.
iovisor/gobpf which is vendored in system-probe is also requiring up-to-date glibc headers. That’s why it is also built in dedicated build environment.

Both libbcc and system-probe are then copied to the final image, in the omnibus tree (/opt/datadog-agent/embedded) where omnibus validates that all symbols found in this library can be resolved in libraries that are also found in the omnibus tree.

This the first part of #4619 which aims at introducing an eBPF-based check that would perform runtime eBPF compilation.
Current implementation of system-probe does a one-time compile-time eBPF code compilation and then do runtime kernel structures offset guessing.
This approach has some limitations that we would like to overcome by doing runtime eBPF code compilation.
iovisor/bcc and its GO wrapper iovisor/gobpf are providing a convenient framework to implement runtime eBPF code compilation.