Unset PYTHONPATH env var (#5001)
Unset PYTHONPATH env var before initializing the embedded Python runtime.

On Windows, it's not uncommon to set PYTHONPATH. If set, it would be picked up by the embedded Python, mixing our bundled site-packages with the system ones (and likely breaking things). This makes this behaviour opt-in through a config parameter.