Make the docker entrypoint script honor DD_CRI_SOCKET_PATH
In order to fix an issue when the docker daemon is restarted while
a containerized agent is connected to it, the docker socket is now
mounted in a different path.

This change was effective in the Helm chart v2.

We must now ensure that the `DD_CRI_SOCKET_PATH` parameter is honored everywhere.

Fixes helm/charts#21223