Do not use DD_CRI_SOCKET_PATH for a docker socket
Setting `DD_CRI_SOCKET_PATH` for a docker socket will enable the `cri` check
whereas the docker daemon doesn't support the CRI API.

Also, when checking the existence of `/var/run/docker.sock`, we must check that
it's really a UNIX socket.
The issue is when the agent is started in a non-docker container with a bind
mount of `/var/run/docker.sock` by mistake whereas there's no docker on the host.
In that case, a new directory named `/var/run/docker.sock/` will be created and
mounted inside the container.
In this case of configuration error, we mustn't activate the `docker` features.