[ebpf] update `tracer-ebpf.go` due to the changes in the builders.
We've noticed that the new builders are not generating exactly the same
`tracer-ebpf.o` output. That may be due to a slight package change in clang
(8.0.1-7 -> 8.0.1-8 even if a bit unlikely as it is a package change).
We've validated that the tests are still passing properly with the new binary and
we're then updating it.