[system-probe] fill in connection direction for UDP connections  (#4869)
* Re-format tracer-ebpf.c

I'm doing this in preparation of a larger eBPF PR which I want to run
the formatter on, but don't want these changes pulled in.

* Add eBPF side of UDP connection direction tracking

This implements a similar scheme that we have had in place for TCP,
where a BPF map is used to keep track of a list of UDP ports which are
listening.

This works by making the following additions:

kprobe and kretprobe for sys_socket (the socket system call): this lets
us keep track of calls to socket() which are of type=datagram and
domain=ipv{4,6}. Since this is a system call with a stable ABI, we can
skip offset guessing.

map for pending_sockets, and unbound_sockets. The former is used to
keep track of state before and after the socket() system call. The
latter is used for remembering which sockets have not yet been bound.

kprobe for sys_bind: this lets us capture the address a socket is being
bound to. Combined with the unbound_sockets map, we can determine when a
UDP server is being created.