Use pkg/forwarder for submitting process-agent payloads
This commit modifies the process-agent to use the pkg/forwarder facility for sending HTTP payloads.  This will ensure that payloads are retried until they are successful.

One difference is that this will submit the API key in a different header than we are currently using.  The intake API will be modified to read this header.

We're also sending a new header, X-DD-Agent-Timestamp that indicates the time that we attempted to send the payload