pkg/trace/writer: fix pooling bug when multiple APM endpoints are configured (#5117)
This change fixes a race condition which occurs in situations where multiple target endpoints are set by creating one copy of the payload for each endpoint. While this may be slower, it's ok because this setting is only used internally as most people only send to one endpoint (US or EU).