Add option to generate flare locally
In a containerized environment such as helm each datadog agent runs in a
separate container (trace, process, ...). They share the same network
with the core agent but not the same filesystem. This means is
impossible to generate flare for the local container (the flare will be
created in the core agent container instead).

This PR add an option to force the CLI to create the archive instead.