Collect network metadata if running docker in host mode (#5218)
gohai, which is responsible for detecting host metadata such as a host's IP address would skip collecting all network data if it detected that the agent was running in a container. This is generally correct, as a container gets a dedicated network namespace, which means functions like net.Interfaces() returns values that only make sense inside the container.

However, if we detect docker0, this generally means we're running the container in host
mode, and we can get at useful network information.

Getting at host IPs means network performance monitoring can resolve traffic destined for these hosts.