[cmd/check] Fix for integrations with both JMX and non-JMX instances  (#5255)
* Revert "Revert "Fix check command when there are JMX and Python/Go instances (#5229)" (#5248)"

This reverts commit cb7a6d87977aea968056c5ce4b3d340fa77c3a9b.

* [cmd/check] Fix when integration has both JMX and non JMX instances

Calls `common.Setup*` (in particular `common.SetupAutoConfig`) are not
idempotent, and are only meant to be called once in the process
lifetime. Otherwise, some components can end up misconfigured, in
particular rtloader/python interpreter.

This affects integrations that both have JMX and non-JMX instances,
for the `check` command which uses some functionality of the `jmx`
command for JMX instances.

To fix this, make sure we only ever call the `common.Setup*` functions
once in the `check` and `jmx` commands in particular.

The only functional change introduced by this commit is the check name
passed to the `check` command should now correctly be taken into
account for JMXFetch checks.

* [jmx] Support 'off' log level

'off' is supported both by the Agent's logger and JMXFetch's logger,
so let's add it.