Use pkg/forwarder for submitting process-agent payloads (#5193)
* Use pkg/forwarder for submitting process-agent payloads

This commit modifies the process-agent to use the pkg/forwarder facility for sending HTTP payloads.  This will ensure that payloads are retried until they are successful.

One difference is that this will submit the API key in a different header than we are currently using.  The intake API will be modified to read this header.

We're also sending a new header, X-DD-Agent-Timestamp that indicates the time that we attempted to send the payload

* Add more documentation around why we're consuming the responses channel

* Use a separate mock endpoint for orchestrator payloads

This will ensure that we test that orchestrator payloads are sent to the orchestrator enpdoint

* Return early if the orchestrator forwarder does not start

* Fix import formatting per agent team style guidelines

* Log an error if we encounter an unknown payload type to submit

* Ensure that the process agent stops if the forwarders cannot be started

* Ensure that response channel is closed and not left dangling

This commit ensures that the response channel is always closed for the caller by reading from a separate internal results channel and waiting for the expected number of responses.  If the number of expected responses does not arrive within a a timeout period, the results channel is closed and the goroutine gives up on waiting

* Ensure that the goroutine reading responses exits after receiving all expected messages