[telemetry+dogstatsd] avoid heap escape in telemetry.Counter (#5251)
The `...string` syntax used in the telemetry pkg escapes the strings on
the heap. Because it is in a hot path, it's important having access to
methods not creating any heap escape.

This PR is adding those methods to the `telemetry.Counter` type. This
will reduce the total garbage memory generated by the Dogstatsd `parseMetricMessage`
and help reduce memory allocation during the spike of metrics sent to Dogstatsd.