Add a retry loop around `manifest-tool push` to workaround notary intermittent failures
The retry loop was introduced for `docker push` with #5040.

We still have some issues with `manifest-tool push` like:
```
Inspect of image "datadog/agent-amd64:7.19.0-rc.8-jmx-win1909" failed with error: manifest unknown: manifest unknown
```

Relaunching the job that produced that error was enough to get it fixed.
So, that was well an intermittent failure.