Limit how much data can be enqueued in memory
This commit modifies the process agent to enforce a strict size limit on the amount of payload bytes that can be enqueued in memory.  By default, we limit this to 75mb worth of payload bytes across process/container/network/pod checks.  If enqueuing a payload for delivery would exceed the maximum size, that payload is dropped.

The delivery queue sizes are also increased to accommodate a much larger set of data.  If the queue size is reached before the size limit, sends will block waiting for space.