[forwarder] Add option to reset connections at configurable interval
Disabled by default. When enabled (with an interval > 0), forces the
forwarder to re-create its connections to Datadog at the configured
interval.

The implementation makes as few assumptions as possible on the behavior
of Go's `http.Client`, and simply re-creates a new client when a reset
is requested.

Also, adds related expvars/telemetry based on the hooks provided by
`httptrace.ClientTrace`, which allow tracking DNS lookups and TCP
connection creation.