Limit how much data can be enqueued in memory (#5477)
* Use separate delivery queues for each intake endpoint

This commit creates separate delivery queues for each intake endpoint.  As of now, this means that there will be a separate queue for pod payloads and one for process/container/rt/network checks.

* Limit how much data can be enqueued in memory

This commit modifies the process agent to enforce a strict size limit on the amount of payload bytes that can be enqueued in memory.  By default, we limit this to 75mb worth of payload bytes across process/container/network/pod checks.  If enqueuing a payload for delivery would exceed the maximum size, that payload is dropped.

The delivery queue sizes are also increased to accommodate a much larger set of data.  If the queue size is reached before the size limit, sends will block waiting for space.

* Introduce WeightedQueue for managing check result data in memory

Using a weighted queue allows us to limit the amount of check data in memory and remove older items when the queue is full

This commit also modifies the timestamp handling to set the timestamp as the time the check data was generated, not the time it was sent. This ensures that the timestamps are properly attributed to the data should it sit in memory for a while

* Add check name to error messages logged in readResponseStatuses

* Log the number of active clients when enabling RT mode

* Fix config initialization

* Document the behavior of Poll()

* Increase the number of in-flight requests allowed in the forwarder

* Revert bad merge.  API key checking should be disabled.

* Fix weight adjustment during item eviction

* Add documentation on how the weighted queue mutex is used

* Add a channel direction to the forwarder response channel

* Sum the active clients across all endpoints