[pkg/metrics/series] Only use device_name if device is not present (#5569)
With the previous behavior, the value of serie.Device could be
overwritten by the contents of the device_name tag if the device_name
tag appears after the device tag in the serie.Tags list. This could
lead to a device value change when upgrading an Agent, as device_name
doesn't necessarily have the same value as device.
To fix this, the device_name tag is only used if no device tag is present.

Moreover, the previous logic removes the device_name tag,
which is not desired.
To fix this, the device_name tag is not removed anymore when populating
the device field.