Allow reporting on cgroup statistics when the container id isn't in all paths
When running a Docker container under systemd you're supposed to pass
the `--cgroup-parent` argument to the run command. This is so systemd is
aware of the processes running inside of the Docker container.

This completely changes the cgroups in use, and some of them no longer
contain the container id.

They go from being `/docker/CID/` to
`/system.slice/ecs-agent.service/CID`, but `memory`, `pids`, `blkio`,
`devices`, and `cpu` are in `/system.slice/ecs-agent.service`. I'm
assuming this is because systemd manages those specific namespaces.

Removing the continue inside this statement allows for the first
discovered container id to be associated with the cgroups that aren't
directly associated with that container id.