[gitlab] Fix release pipeline trigger jobs (#5609)
Make release pipeline trigger jobs always run (they can't be set as manual, and if they get skipped once because of a pipeline failure, they can never be retried).

Fixes the RELEASE_VERSION variable sent to the release management pipeline (it needs the -1). This is not very flexible (eg. if we ever change the revision number), but should be enough for now.