[io] On Windows, add option to send lowercase device tag value
Opt-in. Allows sending the device tag with same case as Agent 5.

Tags are lowercased in the intake, except the `host` and `device` tags,
which is why case is an issue here. Another specificity of the `device`
and `host` tags is that there can't be more than one per timeseries.

Therefore the only way to make this code backwards-compatible with
earlier v6/v7 versions while allowing a non-breaking upgrade from v5
is to add an opt-in option.