Fixes problem wherein uninstall leaves the user home directory around, (#5457)
* Fixes problem wherein uninstall leaves the user home directory around,
which after a sufficient number of uninstall/reinstall will leave the
host unable to create a home directory.

Has the following algorithm
- Asks the OS for the root of the user profile directories (home
directories)
- Looks in that directory only for directory name(s) that include the
ddagent user name.  Must still be a wildcard, because we don't know the
exact directory name (because other directories might be there from
previous installs)
- Only deletes the directory if it has been granted explicit access to
the user we're deleting.  This should prevent deleting of the wrong
directory.

* Review feedback