Change go-bindata version for system-probe
Right now, inv -e system-probe.build will fetch an older version of
go-bindata.

inv -e deps will fetch a newer version of go-bindata (and pin that
version to a tag configured in bootstrap.json.)

The two versions have a similar API, so we should standardize on one
version to use.