pkg/trace/obfuscate: avoid infinite loop on non-ASCII numerals (#5778)
This commit fixes two problems that lead to a hang in the sql parser:
 1. SQLTokenizer.scanNumber, does not return a failure when it fails to scan
    a number.
 2. isDigit returns true for non-ASCII numerals, such as the unicode
    Arabic-Indic digits.

IsDigit reporting true for non-ASCII numerals caused scanNumber to be
called on a number that it could not parse. ScanNumber would return
successfully without advancing the stream, and the loop would continue in
the same place, causing an infinite loop.

This commit addresses both issues. IsDigit will now return true only for
ascii numerals 0-9, and scanNumber will return a LexError if it fails to
parse a number.